JCAPTCHA-14 : Correction of sonar issues.
git-svn-id: http://dev.lutece.paris.fr/svn/lutece/portal/trunk/plugins/cms/module-captcha-jcaptcha@49186 bab10101-e421-0410-a517-8ce0973de3ef

+++ src/java/fr/paris/lutece/plugins/captcha/modules/jcaptcha/service/sound/filter/EchoFilter.java
@@ -61,7 +61,7 @@ public class EchoFilter extends SoundFilter
      */
     public EchoFilter( float echoDelay, float decay )
     {
-        int numSampleDelay = ( Math.round( LuteceWordToSound.getSoundsSampleRate(  ) * ( (float) ( echoDelay / 1000 ) ) ) +
+        int numSampleDelay = ( Math.round( LuteceWordToSound.getSoundsSampleRate( ) * ( echoDelay / 1000 ) ) +
             1 );
         _delayBuffer = new short[numSampleDelay];
         _decay = decay / 100;
